Avoid integer overflow in year 2038
See https://en.wikipedia.org/wiki/Year_2038_problem

This patch was done while reviewing potential year-2038 issues in openSUSE.

+++ src/conn.h
@@ -100,7 +100,7 @@ typedef struct {
 	tcp_t *tcp;
 	bool enabled;
 	bool supported;
-	int last_transfer;
+	long long last_transfer;
 	char *message;
 	char *local_if;
